+++ gplugins/klayout/dataprep/__init__.py
@@ -1,0 +1,19 @@
+from gplugins.klayout.dataprep import regions
+from gplugins.klayout.dataprep.regions import (
+    Region,
+    RegionCollection,
+    boolean_not,
+    boolean_or,
+    copy,
+    size,
+)
+
+__all__ = [
+    "Region",
+    "RegionCollection",
+    "boolean_not",
+    "boolean_or",
+    "copy",
+    "regions",
+    "size",
+]