+++ gplugins/tidy3d/modes.py
@@ -22,16 +22,33 @@ import pydantic
 import tidy3d as td
 import xarray
 from gdsfactory.config import logger
-from gdsfactory.pdk import MaterialSpec, get_modes_path
-from gdsfactory.serialization import clean_value_name
+from gdsfactory.pdk import get_modes_path
 from gdsfactory.typings import PathType
+from pydantic import BaseModel
 from tidy3d.plugins import waveguide
 from tqdm.auto import tqdm
 
-from gplugins.tidy3d.materials import get_medium
+from gplugins.tidy3d.materials import MaterialSpecTidy3d, get_medium
 
 Precision = Literal["single", "double"]
 nm = 1e-3
+
+
+def custom_serializer(data: str | float | BaseModel) -> str:
+    # If data is a string, just return it.
+    if isinstance(data, str | None | np.ndarray):
+        return data
+
+    # If data is a float, convert it to a string.
+    if isinstance(data, float | int):
+        return str(data)
+
+    # If data is an instance of Pydantic's BaseModel, serialize it to JSON.
+    if isinstance(data, BaseModel):
+        return data.json()
+
+    # For all other data types, raise an exception.
+    raise ValueError(f"Unsupported data type: {type(data)}")
 
 
 class Waveguide(pydantic.BaseModel):
@@ -47,6 +64,7 @@ class Waveguide(pydantic.BaseModel):
             - string: material name.
             - float: refractive index.
             - float, float: refractive index real and imaginary part.
+            - td.Medium: tidy3d medium.
             - function: function of wavelength.
         clad_material: top cladding material.
         box_material: bottom cladding material.
@@ -107,9 +125,9 @@ class Waveguide(pydantic.BaseModel):
     wavelength: float | Sequence[float] | Any
     core_width: float
     core_thickness: float
-    core_material: MaterialSpec | td.CustomMedium
-    clad_material: MaterialSpec
-    box_material: MaterialSpec | None = None
+    core_material: MaterialSpecTidy3d
+    clad_material: MaterialSpecTidy3d
+    box_material: MaterialSpecTidy3d | None = None
     slab_thickness: float = 0.0
     clad_thickness: float | None = None
     box_thickness: float | None = None
@@ -154,10 +172,9 @@ class Waveguide(pydantic.BaseModel):
         cache_path.mkdir(exist_ok=True, parents=True)
 
         settings = [
-            f"{setting}={clean_value_name(getattr(self, setting))}"
+            f"{setting}={custom_serializer(getattr(self, setting))}"
             for setting in sorted(self.__fields__.keys())
         ]
-
         named_args_string = "_".join(settings)
         h = hashlib.md5(named_args_string.encode()).hexdigest()[:16]
         return cache_path / f"{self.__class__.__name__}_{h}.npz"
@@ -169,12 +186,23 @@ class Waveguide(pydantic.BaseModel):
         #         or isinstance(self.core_material, td.CustomMedium)):
         if not hasattr(self, "_waveguide"):
             # To include a dn -> custom medium
-            if isinstance(self.core_material, td.CustomMedium):
+            if isinstance(self.core_material, td.CustomMedium | td.Medium):
                 core_medium = self.core_material
             else:
                 core_medium = get_medium(self.core_material)
-            clad_medium = get_medium(self.clad_material)
-            box_medium = get_medium(self.box_material) if self.box_material else None
+
+            if isinstance(self.clad_material, td.CustomMedium | td.Medium):
+                clad_medium = self.clad_material
+            else:
+                clad_medium = get_medium(self.clad_material)
+
+            if self.box_material:
+                if isinstance(self.box_material, td.CustomMedium | td.Medium):
+                    box_medium = self.box_material
+                else:
+                    box_medium = get_medium(self.box_material)
+            else:
+                box_medium = None
 
             freq0 = td.C_0 / np.mean(self.wavelength)
             n_core = core_medium.eps_model(freq0) ** 0.5
@@ -235,11 +263,10 @@ class Waveguide(pydantic.BaseModel):
         """Mode data for this waveguide (cached if cache is enabled)."""
         if not hasattr(self, "_cached_data"):
             filepath = self.filepath
-            if filepath and filepath.exists():
-                if not self.overwrite:
-                    logger.info(f"load data from {filepath}.")
-                    self._cached_data = np.load(filepath)
-                    return self._cached_data
+            if filepath and filepath.exists() and not self.overwrite:
+                logger.info(f"load data from {filepath}.")
+                self._cached_data = np.load(filepath)
+                return self._cached_data
 
             wg = self.waveguide
 
@@ -361,7 +388,7 @@ class Waveguide(pydantic.BaseModel):
         field_name: str,
         value: str = "real",
         mode_index: int = 0,
-        wavelength: float = None,
+        wavelength: float | None = None,
         **kwargs,
     ) -> None:
         """Plot the selected field distribution from a waveguide mode.
@@ -375,6 +402,11 @@ class Waveguide(pydantic.BaseModel):
             kwargs: keyword arguments passed to xarray.DataArray.plot.
         """
         data = self._data[field_name]
+
+        if mode_index >= self.num_modes:
+            raise ValueError(
+                f"mode_index = {mode_index} must be less than num_modes {self.num_modes}"
+            )
 
         if self.num_modes > 1:
             data = data[..., mode_index]
@@ -404,6 +436,10 @@ class Waveguide(pydantic.BaseModel):
         data_array = xarray.DataArray(
             data.T, coords={"y": self._data["y"], "x": self._data["x"]}
         )
+
+        if value == "dB":
+            kwargs.update(vmin=-20)
+
         data_array.name = field_name
         artist = data_array.plot(**kwargs)
         artist.axes.set_aspect("equal")
@@ -416,7 +452,10 @@ class Waveguide(pydantic.BaseModel):
         """Show waveguide representation."""
         return (
             f"{self.__class__.__name__}("
-            + ", ".join(f"{k}={getattr(self, k)!r}" for k in self.__fields__.keys())
+            + ", ".join(
+                f"{k}={custom_serializer(getattr(self, k))!r}"
+                for k in self.__fields__.keys()
+            )
             + ")"
         )
 
@@ -965,18 +1004,26 @@ if __name__ == "__main__":
     #     overwrite=True
     # )
 
+    import matplotlib.pyplot as plt
+
     strip = Waveguide(
         wavelength=1.55,
         core_width=1.0,
         slab_thickness=0.0,
-        core_material="si",
+        # core_material="si",
+        core_material=td.material_library["cSi"]["Li1993_293K"],
         clad_material="sio2",
         core_thickness=220 * nm,
         num_modes=4,
     )
-    w = np.linspace(400 * nm, 1000 * nm, 7)
-    n_eff = sweep_n_eff(strip, core_width=w)
-    fraction_te = sweep_fraction_te(strip, core_width=w)
+    # strip._data
+    # strip.filepath
+    # strip.plot_index()
+    strip.plot_field("Ex", mode_index=0, wavelength=1.55, value="dB")
+    plt.show()
+    # w = np.linspace(400 * nm, 1000 * nm, 7)
+    # n_eff = sweep_n_eff(strip, core_width=w)
+    # fraction_te = sweep_fraction_te(strip, core_width=w)
 
     # t = np.linspace(0.2, 0.25, 6)
     # w = np.linspace(0.4, 0.6, 5)