+++ docs/changelog.md
@@ -1,6 +1,17 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased](https://github.com/gdsfactory/gdsfactory/compare/v0.0.2...main)
+## [Unreleased](https://github.com/gdsfactory/gplugins/compare/v0.1.0...main)
+
+## [0.1.1](https://github.com/gdsfactory/gplugins/compare/v0.1.1...v0.1.0)
+
+- move klayout/dataprep from gdsfactory into gplugins/klayout. Add tests.
+- improve tidy3d plugin
+    - Enable sidewall angles in FDTD simulations
+    - improve tidy3d mode solver and tests [PR](https://github.com/gdsfactory/gplugins/pull/25)
+
+## [0.1.0](https://github.com/gdsfactory/gplugins/compare/v0.1.0...v0.0.3)
+
+- add support for tidy3d materials [PR](https://github.com/gdsfactory/gplugins/pull/17)
 
 ## [0.0.3](https://github.com/gdsfactory/gplugins/compare/v0.0.3...v0.0.2)